+++ src/kiwano/ui/Menu.h
@@ -38,7 +38,7 @@ namespace kiwano
 		bool IsEnable() const;
 
 		// 获取菜单中的按钮数量
-		size_t GetButtonCount() const;
+		UInt32 GetButtonCount() const;
 
 		// 设置菜单启用或禁用
 		void SetEnabled(
